+++ src/Sites/UI/views/site-selection/site-selection.blade.php
@@ -1,10 +1,8 @@
-@php
-    $sites = $this->getSites();
-@endphp
-
 <div>
     <div class="flex items-center gap-2">
-        <livewire:chief-wire::model-site-toggle :model="$this->model" />
+        @if($this->isAllowedToSelectSites() || count($this->getSites()) > 1)
+            <livewire:chief-wire::model-site-toggle :model="$this->model" />
+        @endif
 
         @if($this->isAllowedToSelectSites())
             <x-chief::button wire:click="edit" size="sm" variant="grey" title="Sites aanpassen" class="shrink-0">